% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_barplot.R
\name{bc_barplot}
\alias{bc_barplot}
\title{(B)irth (C)ohort Coverage Bar Plot}
\usage{
bc_barplot(
  data,
  vaccine,
  birth_cohorts = NULL,
  within_ADM1 = NULL,
  palette = "YlOrBr"
)
}
\arguments{
\item{data}{Data frame. The output from the \code{pahoabc::bc_coverage} function.}

\item{vaccine}{Character. A string specifying the vaccine of interest. Only one vaccine.}

\item{birth_cohorts}{Numeric (optional). A vector specifying the birth cohort(s) for which coverage should be plotted. If \code{NULL} (default), all years are plotted.}

\item{within_ADM1}{Character (optional). When analyzing data at the "ADM2" level, this optional character vector lets you specify one or several "ADM1" to filter. Default is \code{NULL}, which means no filtering by "ADM1".}

\item{palette}{Character (optional). A RColorBrewer palette name for the cohort colors. Default is \code{"YlOrBr"}.}
}
\value{
A ggplot object representing the bar plot.
}
\description{
Generates a bar plot of birth cohort coverage, grouped by geographic area and colored by birth cohort year.
}
