% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_barplot.R
\name{cs_barplot}
\alias{cs_barplot}
\title{(C)omplete (S)chedule Coverage Bar Plot}
\usage{
cs_barplot(data, birth_cohorts = NULL)
}
\arguments{
\item{data}{The output from the \code{pahoabc::cs_coverage} function.}

\item{birth_cohorts}{Numeric (optional). A vector specifying the birth cohort(s) for which coverage should be plotted. If \code{NULL} (default), all years are plotted.}
}
\value{
A ggplot object representing the bar plot.
}
\description{
Generates a bar plot of complete schedule vaccination coverage, grouped by geographic area and colored by birth cohort.
}
