% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_upset.R
\name{cs_upset}
\alias{cs_upset}
\title{(C)omplete (S)chedule Coverage Upset Plot}
\usage{
cs_upset(
  data.EIR,
  data.schedule,
  birth_cohort,
  denominator = NULL,
  min_size = 1
)
}
\arguments{
\item{data.EIR}{A data frame containing individual vaccination records. See \code{pahoabc.EIR} for expected structure.}

\item{data.schedule}{A data frame defining the vaccination schedule. See \code{pahoabc.schedule} for expected structure.}

\item{birth_cohort}{Numeric. A single birth year for which to calculate and visualize coverage.}

\item{denominator}{The denominator to use. If \code{NULL} (default), then the number of unique IDs in \code{data.EIR} is used.}

\item{min_size}{The minimum number of doses (as a percentage of the \code{denominator}) a group has to have in order to be shown in the plot. Default is 1 percent.}
}
\value{
A ComplexUpset plot.
}
\description{
Generates an UpSet plot showing the number of individuals who received various combinations of vaccine doses defined in a vaccination schedule, for a given birth cohort.
Individuals with a complete schedule are highlighted.
}
