% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_upsetplot.R
\name{cs_upsetplot}
\alias{cs_upsetplot}
\title{(C)omplete (S)chedule Coverage Upset Plot}
\usage{
cs_upsetplot(
  data.EIR,
  data.schedule,
  birth_cohort,
  denominator = NULL,
  min_size = 1,
  set_order = NULL
)
}
\arguments{
\item{data.EIR}{A data frame containing individual vaccination records. See \code{pahoabc.EIR} for expected structure.}

\item{data.schedule}{A data frame defining the vaccination schedule. See \code{pahoabc.schedule} for expected structure.}

\item{birth_cohort}{Numeric. A single birth year for which to calculate and visualize coverage.}

\item{denominator}{The denominator to use. If \code{NULL} (default), then the number of unique IDs in \code{data.EIR} is used.}

\item{min_size}{The minimum number of doses (as a percentage of the \code{denominator}) a group has to have in order to be shown in the plot. Default is 1 percent.}

\item{set_order}{A vector containing the desired order of the doses as they will appear in the combination matrix. They must match the names in the \code{dose} column of \code{data.EIR} and \code{data.schedule}. If NULL, uses the default ordering scheme of \code{ComplexUpset::upset()}.}
}
\value{
A ComplexUpset plot.
}
\description{
Generates an UpSet plot showing the number of individuals who received various combinations of vaccine doses defined in a vaccination schedule, for a given birth cohort.
Individuals with a complete schedule are highlighted.
}
