% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_barplot.R
\name{dcc_barplot}
\alias{dcc_barplot}
\title{(D)ate (C)onsistency (C)omparison Barplot}
\usage{
dcc_barplot(
  data,
  date_1_name = "date 1",
  date_2_name = "date 2",
  within_ADM1 = NULL,
  plot_missing = TRUE
)
}
\arguments{
\item{data}{The output from the \code{pahoabc::dcc_rate} function.}

\item{date_1_name}{Character. Label for the first date variable, used in the plot subtitle.}

\item{date_2_name}{Character. Label for the second date variable, used in the plot subtitle.}

\item{within_ADM1}{Character (optional). When analyzing data at the "ADM2" level, this optional character vector lets you specify one or several "ADM1" to filter. Default is \code{NULL}, which means no filtering by "ADM1".}

\item{plot_missing}{Boolean (default TRUE). Plots DATE_MISSING records as its own category. Defaults to TRUE so all bars add to 100}
}
\value{
A ggplot object representing the bar plot.
}
\description{
Generates a bar plot of date consistencies, grouped by geographic area and colored by consistency result.
}
