% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropoutplot.R
\name{dropoutplot}
\alias{dropoutplot}
\title{Nominal dropout rate}
\usage{
dropoutplot(nominal_dropout_data, vertical = F, order = "alpha")
}
\arguments{
\item{nominal_dropout_data}{Dataframe formatted from the result of the nominal_dropout function}

\item{vertical}{Boolean that shows the plot vertical or horizontal, default is F which corresponds to horizontal. True corresponds to vertical.}

\item{order}{Organizes the bars based on three options: "alpha", "desc" or "asc", default is "alpha"}
}
\value{
A ggplot simple barplot for the administrative level used to calculate the nominal dropout. Includes the nominal dropout rate, numerator, denominator and adherence rate. based on the administrative level, birth cohort and vaccines selected
}
\description{
Nominal dropout rate
}
