% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nod_barplot.R
\name{nod_barplot}
\alias{nod_barplot}
\title{Nominal dropout plot}
\usage{
nod_barplot(data, order = "alpha", within_ADM1 = NULL)
}
\arguments{
\item{data}{Dataframe formatted from the output of the \code{pahoabc::nod_dropout} function}

\item{order}{Organizes the bars based on three options: "alpha", "desc" or "asc", default is "alpha"}

\item{within_ADM1}{When analyzing data at the "ADM2" level, this optional character vector lets you specify one or several "ADM1" to filter. Default is \code{NULL}, which means no filtering by "ADM1".}
}
\value{
A ggplot simple barplot for the administrative level used to calculate the nominal dropout. Includes the nominal dropout rate, numerator, denominator and adherence rate. based on the administrative level, birth cohort and vaccines selected
}
\description{
Nominal dropout plot
}
