% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominal_dropout.R
\name{nominal_dropout}
\alias{nominal_dropout}
\title{Nominal dropout rate}
\usage{
nominal_dropout(
  data = pahoabc.EIR,
  vac_init,
  vac_end,
  geo_level = NA,
  birth_cohort = NA
)
}
\arguments{
\item{data}{Nominal data frame}

\item{vac_init}{Initial vaccine to use in nominal dropout calculation}

\item{vac_end}{Final vaccine to use in nominal dropout calculation}

\item{geo_level}{Administrative level to group by for calculation, can be "ADM1" or "ADM2". If no geo_level is specified the default is NA which represents ADM0}

\item{birth_cohort}{Birth cohort to calculate for.}
}
\value{
A data frame with the calculated coverage for the administrative level used to calculate. Includes the nominal dropout rate, numerator, denominator and adherence rate. based on the administrative level, birth cohort and vaccines selected
}
\description{
Nominal dropout rate
}
