% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pahoabc.EIR.R
\docType{data}
\name{pahoabc.EIR}
\alias{pahoabc.EIR}
\title{PAHOABC Electronic Immunization Registry (EIR)}
\format{
A data frame with X rows and 8 variables:
\describe{
  \item{ID}{A unique identifier for each person.}
  \item{birth_date}{The person's birth date.}
  \item{vaccination_date}{The date of vaccination.}
  \item{ADM1_residence}{The first geographic administrative level of the person's residence.}
  \item{ADM2_residence}{The second geographic administrative level of the person's residence.}
  \item{ADM1_occurrence}{The first geographic administrative level where the vaccination occurred.}
  \item{ADM2_occurrence}{The second geographic administrative level where the vaccination occurred.}
  \item{dose}{The dose received during the vaccination event.}
}
}
\source{
pahoabc
}
\usage{
pahoabc.EIR
}
\description{
A dataset containing nominal vaccination events from an immunization information system.
Each row represents a vaccination act for a specific person, along with relevant geographic and demographic information.
}
\keyword{datasets}
