% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pahoabc.pop.ADM0.R
\docType{data}
\name{pahoabc.pop.ADM0}
\alias{pahoabc.pop.ADM0}
\title{PAHOABC Aggregated Population Data by ADM0}
\format{
A data frame with X rows and 3 variables:
\describe{
  \item{year}{The year of the population data.}
  \item{age}{The age (in years) of the population for that year and ADM1.}
  \item{population}{The population for that year and ADM1.}
}
}
\source{
pahoabc
}
\usage{
pahoabc.pop.ADM0
}
\description{
A dataset containing aggregated population information for the whole country (ADM0).
Each row provides the population of children aged 0 and 1 years for a given year.
}
\keyword{datasets}
