% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pahoabc.pop.ADM1.R
\docType{data}
\name{pahoabc.pop.ADM1}
\alias{pahoabc.pop.ADM1}
\title{PAHOABC Aggregated Population Data by ADM1}
\format{
A data frame with X rows and 4 variables:
\describe{
  \item{ADM1}{The name of the first geographic administrative level.}
  \item{year}{The year of the population data.}
  \item{pop_0}{The population of children aged 0 years.}
  \item{pop_1}{The population of children aged 1 year.}
}
}
\source{
pahoabc
}
\usage{
pahoabc.pop.ADM1
}
\description{
A dataset containing aggregated population information by the first geographic administrative level (ADM1).
Each row provides the population of children aged 0 and 1 years for a given year and administrative region.
}
\keyword{datasets}
