% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_barplot.R
\name{roc_barplot}
\alias{roc_barplot}
\title{(R)esidence (Oc)currence Bar Plot}
\usage{
roc_barplot(data, vaccine, within_ADM1 = NA)
}
\arguments{
\item{data}{The output from the `pahoabc::roc_coverage()` function.}

\item{vaccine}{A character string specifying the vaccine of interest. Only one vaccine.}

\item{within_ADM1}{When analyzing data at the `ADM2` level, this optional parameter lets you specify a particular `ADM1` to filter. Default is `NA`, which means no filtering by `ADM1`.}
}
\value{
A ggplot object representing the bar plot.
}
\description{
This function creates a bar plot showing vaccination coverage by place of residence and place of vaccination occurrence.
}
