% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_coverage_by.R
\name{roc_coverage_by}
\alias{roc_coverage_by}
\title{(R)esidence or (Oc)currence Coverage}
\usage{
roc_coverage_by(
  coverage_type,
  data.EIR,
  data.schedule,
  data.pop,
  geo_level,
  years = NULL,
  vaccines = NULL
)
}
\arguments{
\item{coverage_type}{A character string specifying the type of analysis: either "residence" or "occurrence".}

\item{data.EIR}{A data frame containing individual vaccination records. See \code{pahoabc.EIR} for expected structure.}

\item{data.schedule}{A data frame defining the vaccination schedule. See \code{pahoabc.schedule} for expected structure.}

\item{data.pop}{A data frame with population denominators. See \code{pahoabc.pop.ADMX} for structure examples.}

\item{geo_level}{The geographic level to aggregate results by. Must be "ADM0", "ADM1" or "ADM2". \code{data.pop} must contain the columns to match.}

\item{years}{Numeric (optional). The years for which the coverage calculation is done. If \code{NULL} (default), all vaccination years in \code{data.EIR} are included.}

\item{vaccines}{Character (optional). A character vector specifying the doses to include in the analysis. If \code{NULL} (default), all vaccines in \code{data.EIR} are included.}
}
\value{
A data frame with calculated coverage for the specified analysis type, year, vaccines, and geographic level.
}
\description{
Calculates vaccine coverage by either residence or occurrence at a specified geographic level, for a set of years and vaccines.
}
