% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_heatmap.R
\name{roc_heatmap}
\alias{roc_heatmap}
\title{(R)esidence (Oc)currence Heatmap}
\usage{
roc_heatmap(data)
}
\arguments{
\item{data}{The output from the `pahoabc::roc_distribution` function.}
}
\value{
A ggplot object representing the heatmap.
}
\description{
This function generates a heatmap displaying the proportion of doses administered
by place of occurrence, for each place of residence.
}
