% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocplot.R
\name{rocplot}
\alias{rocplot}
\title{(R)esidence (O)ccurrence (C)overage Plot}
\usage{
rocplot(data, vaccine, within_ADM1 = NA)
}
\arguments{
\item{data}{The output from the `pahoabc::roc()` function. This data should include columns for `dose`, `coverage_type`, `coverage`, `year_vax`, and geographic variables (`ADM1` and optionally `ADM2`).}

\item{vaccine}{A character string specifying the vaccine of interest. Only one vaccine.}

\item{within_ADM1}{When analyzing data at the `ADM2` level, this optional parameter lets you specify a particular ADM1 to filter. Default is `NA`, which means no filtering by `ADM1`.}
}
\value{
A ggplot object representing the bar plot.
}
\description{
This function creates a bar plot showing vaccination coverage by the place of residence and the place of vaccination occurrence, grouped by a geographic level.
}
