% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_barplot.R
\name{cs_barplot}
\alias{cs_barplot}
\title{(C)omplete (S)chedule Coverage Bar Plot}
\usage{
cs_barplot(data, birth_cohorts = NULL, within_ADM1 = NULL)
}
\arguments{
\item{data}{The output from the \code{pahoabc::cs_coverage} function.}

\item{birth_cohorts}{Numeric (optional). A vector specifying the birth cohort(s) for which coverage should be plotted. If \code{NULL} (default), all years are plotted.}

\item{within_ADM1}{Character (optional). When analyzing data at the "ADM2" level, this optional character vector lets you specify one or several "ADM1" to filter. Default is \code{NULL}, which means no filtering by "ADM1".}
}
\value{
A ggplot object representing the bar plot.
}
\description{
Generates a bar plot of complete schedule vaccination coverage, grouped by geographic area and colored by birth cohort.
}
