% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_upset.R
\name{cs_upset}
\alias{cs_upset}
\title{(C)omplete (S)chedule Coverage Upset Plot}
\usage{
cs_upset(data.EIR, data.pop, data.schedule, birth_cohort)
}
\arguments{
\item{data.EIR}{A data frame containing individual vaccination records. See \code{pahoabc.EIR} for expected structure.}

\item{data.pop}{A data frame with population denominators. Only works at ADM0. Must follow structure in \code{pahoabc.pop.ADM0}.}

\item{data.schedule}{A data frame defining the vaccination schedule. See \code{pahoabc.schedule} for expected structure.}

\item{birth_cohort}{Numeric. A single birth year for which to calculate and visualize coverage.}
}
\value{
A ComplexUpset plot.
}
\description{
Generates an UpSet plot showing the number of individuals who received various combinations of vaccine doses defined in a vaccination schedule, for a given birth cohort.
Individuals with a complete schedule are highlighted.
}
