% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_inconsistent.R
\name{dcc_inconsistent}
\alias{dcc_inconsistent}
\title{(D)ate (C)onsistency (C)omparison Inconsistent Records}
\usage{
dcc_inconsistent(data.EIR, date_1, date_2, birth_cohorts = NULL)
}
\arguments{
\item{data.EIR}{Data frame. A data frame containing individual vaccination records. See \code{pahoabc.EIR} for expected structure.}

\item{date_1}{Character. The name of a DATE formatted variable present in the EIR. This variable is the date we are checking for consistency.}

\item{date_2}{Character. The name of a DATE formatted variable present in the EIR. This variable is the date we are checking consistency against. Represents a date chronologically earlier than date_1.}

\item{birth_cohorts}{Numeric (optional). A vector specifying the birth cohort(s) for which inconsistent records should be returned. If \code{NULL} (default), all available cohorts are used.}
}
\value{
A data frame containing the inconsistent records in the database and the time diferential between them.
}
\description{
This function returns a record-level table of all inconsistent and "date missing" entries, including the day difference between the two dates.
}
