% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_rate.R
\name{dcc_rate}
\alias{dcc_rate}
\title{(D)ate (C)onsistency (C)omparison Rate}
\usage{
dcc_rate(data.EIR, date_1, date_2, geo_level = "ADM0")
}
\arguments{
\item{data.EIR}{Data frame. A data frame containing individual vaccination records. See \code{pahoabc.EIR} for expected structure.}

\item{date_1}{Character. The name of a DATE formatted variable present in the EIR. This variable is the date we are checking for consistency.}

\item{date_2}{Character. The name of a DATE formatted variable present in the EIR. This variable is the date we are checking consistency against. Represents a date chronologically earlier than date_1.}

\item{geo_level}{Character. The geographic level to aggregate results by. Must be "ADM0", "ADM1" or "ADM2". If not specified, the default is "ADM0".}
}
\value{
A data frame containing the consistency rate and inconsistency rate (inverse proportion) in percentages for the specified \code{geo_level}.
}
\description{
This function calculates the consistency comparison rate between two date variables, further disagreggating by geographic level, using data from the electronic immunization registry (EIR).
}
