% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_barplot.R
\name{roc_barplot}
\alias{roc_barplot}
\title{(R)esidence (Oc)currence Bar Plot}
\usage{
roc_barplot(data, year, vaccine, within_ADM1 = NULL)
}
\arguments{
\item{data}{The output from the \code{pahoabc::roc_coverage()} function.}

\item{year}{A numeric value specifying the year of the data to be plotted. Only one year.}

\item{vaccine}{A string specifying the vaccine of interest. Only one vaccine.}

\item{within_ADM1}{When analyzing data at the "ADM2" level, this optional string lets you specify a particular "ADM1" to filter. Default is \code{NULL}, which means no filtering by "ADM1".}
}
\value{
A ggplot object representing the bar plot.
}
\description{
This function creates a bar plot showing vaccination coverage by place of residence and place of vaccination occurrence.
}
