% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_heatmap.R
\name{roc_heatmap}
\alias{roc_heatmap}
\title{(R)esidence (Oc)currence Heatmap}
\usage{
roc_heatmap(data, digits = 0)
}
\arguments{
\item{data}{The output from the \code{pahoabc::roc_distribution} function.}

\item{digits}{The number of digits to show in the labels. Default is 0.}
}
\value{
A ggplot object representing the heatmap.
}
\description{
This function generates a heatmap displaying the proportion of doses administered
by place of occurrence, for each place of residence.
}
